update fork to actual state

+++ manifests/installasm.pp
@@ -55,7 +55,7 @@
 # @param storage_option
 #
 define oradb::installasm(
-  Enum['11.2.0.4','12.1.0.1','12.1.0.2','12.2.0.1'] $version                       = undef,
+  Enum['11.2.0.4','12.1.0.1','12.1.0.2','12.2.0.1','18.0.0.0','19.0.0.0'] $version = undef,
   String $file                                                                     = undef,
   Enum['HA_CONFIG', 'CRS_CONFIG', 'UPGRADE', 'CRS_SWONLY'] $grid_type              = 'HA_CONFIG',
   Boolean $stand_alone                                                             = true, # in case of 'CRS_SWONLY' and used as stand alone or in RAC
@@ -163,7 +163,7 @@ define oradb::installasm(
 
     if ( $zip_extract ) {
       # In $download_dir, will Puppet extract the ZIP files or is this a pre-extracted directory structure.
-      if ( $version in ['12.2.0.1']) {
+      if ( $version in ['12.2.0.1','18.0.0.0','19.0.0.0']) {
         $file1 = "${file}.zip"
         $total_files = 1
       }
@@ -208,27 +208,28 @@ define oradb::installasm(
         $source = $puppet_download_mnt_point
       }
 
-      if ($version == '12.2.0.1') {
-        exec { "make ${grid_home}":
-          command   => "mkdir -p ${grid_home}",
-          timeout   => 0,
-          logoutput => false,
-          path      => $exec_path,
-          user      => $user,
-          group     => $group,
-          creates   => $grid_home,
-          require   => Db_directory_structure["grid structure ${version}"],
-          before    => Exec["install oracle grid ${title}"],
-        }
-        file { $grid_home:
-          ensure  => directory,
-          recurse => false,
-          replace => false,
-          mode    => '0775',
-          owner   => $user,
-          group   => $group_install,
-          require => [Db_directory_structure["grid structure ${version}"], Exec["make ${grid_home}"]],
-        }
+      if ($version in ['12.2.0.1','18.0.0.0','19.0.0.0']) {
+        # exec { "make ${grid_home}":
+        #   command   => "mkdir -p ${grid_home}",
+        #   timeout   => 0,
+        #   logoutput => false,
+        #   path      => $exec_path,
+        #   user      => $user,
+        #   group     => $group,
+        #   creates   => $grid_home,
+        #   require   => Db_directory_structure["grid structure ${version}"],
+        #   before    => Exec["install oracle grid ${title}"],
+        # }
+        # replaced to the end, else this is double it will be root as owner
+        # file { $grid_home:
+        #   ensure  => directory,
+        #   recurse => true,
+        #   replace => false,
+        #   mode    => '0775',
+        #   owner   => $user,
+        #   group   => $group_install,
+        #   require => [Db_directory_structure["grid structure ${version}"], Exec["make ${grid_home}"]],
+        # }
         exec { "extract ${download_dir}/${file1}":
           command   => "unzip -o ${source}/${file1} -d ${grid_home}",
           timeout   => 0,
@@ -237,7 +238,8 @@ define oradb::installasm(
           user      => $user,
           group     => $group,
           creates   => "${grid_home}/bin",
-          require   => [Db_directory_structure["grid structure ${version}"], File[$grid_home]],
+          require   => Db_directory_structure["grid structure ${version}"],
+          # require   => [Db_directory_structure["grid structure ${version}"], File[$grid_home]],
           before    => Exec["install oracle grid ${title}"],
         }
       } else {
@@ -305,7 +307,7 @@ define oradb::installasm(
       }
     }
 
-    if ($version == '12.2.0.1') {
+    if ($version in ['12.2.0.1','18.0.0.0','19.0.0.0']) {
       $command = "/bin/sh -c 'unset DISPLAY;cd ${grid_home};./gridSetup.sh -silent -waitforcompletion -skipPrereqs -responseFile ${download_dir}/grid_install_${version}.rsp'"
     } else {
       $command = "/bin/sh -c 'unset DISPLAY;cd ${grid_base};${download_dir}/${file_without_ext}/grid/runInstaller -silent -waitforcompletion -ignoreSysPrereqs -ignorePrereq -responseFile ${download_dir}/grid_install_${version}.rsp'"
@@ -374,7 +376,7 @@ define oradb::installasm(
       require   => Exec["install oracle grid ${title}"],
     }
 
-    if ( $version == '12.2.0.1' and $grid_type != 'CRS_SWONLY' ) {
+    if ( $version in ['12.2.0.1','18.0.0.0','19.0.0.0'] and $grid_type != 'CRS_SWONLY' ) {
       exec { "configure asm ${title}":
         timeout   => 0,
         command   => "${grid_home}/gridSetup.sh -executeConfigTools -silent -responseFile ${download_dir}/grid_install_${version}.rsp",
@@ -400,7 +402,7 @@ define oradb::installasm(
     }
     # cleanup
     if ( $zip_extract ) {
-      if ($version != '12.2.0.1') {
+      if ($version != '12.2.0.1' and $version != '18.0.0.0' and $version != '19.0.0.0') {
         exec { "remove oracle asm extract folder ${title}":
           command => "rm -rf ${download_dir}/${file_without_ext}",
           user    => 'root',
